Move test helper and finish test fetch pending requests
